refactor(IDE-15): integrate LSP into existing MonacoWebView
- Remove duplicate MonacoLSPWebView
- Add enableLSP parameter to MonacoWebView
- EditorView uses MonacoWebView with enableLSP for Java files
- Add monacoLSPHTML fallback with light theme
