+++ Sources/Zero/Views/MonacoWebView.swift
@@ -5,6 +5,8 @@ struct MonacoWebView: NSViewRepresentable {
     @Binding var content: String
     var language: String
     var onReady: (() -> Void)?
+    var onCursorChange: ((Int, Int) -> Void)?
+    var enableLSP: Bool = false
     
     func makeNSView(context: Context) -> WKWebView {
         let config = WKWebViewConfiguration()
@@ -15,13 +17,14 @@ struct MonacoWebView: NSViewRepresentable {
         webView.navigationDelegate = context.coordinator
         context.coordinator.webView = webView
         
-        // Load Monaco HTML
-        if let htmlPath = Bundle.main.path(forResource: "monaco", ofType: "html"),
+        // Load Monaco HTML (LSP or standard)
+        let htmlResource = enableLSP ? "monaco-lsp" : "monaco"
+        if let htmlPath = Bundle.main.path(forResource: htmlResource, ofType: "html"),
            let htmlContent = try? String(contentsOfFile: htmlPath) {
             webView.loadHTMLString(htmlContent, baseURL: URL(string: "https://cdnjs.cloudflare.com"))
         } else {
             // Fallback: Load from embedded string
-            let html = Self.monacoHTML
+            let html = enableLSP ? Self.monacoLSPHTML : Self.monacoHTML
             webView.loadHTMLString(html, baseURL: URL(string: "https://cdnjs.cloudflare.com"))
         }
         
@@ -66,6 +69,117 @@ struct MonacoWebView: NSViewRepresentable {
             }
         }
     }
+    
+    // Embedded Monaco LSP HTML as fallback
+    static let monacoLSPHTML = """
+    <!DOCTYPE html>
+    <html>
+    <head>
+        <meta charset="UTF-8">
+        <style>
+            * { margin: 0; padding: 0; box-sizing: border-box; }
+            html, body { height: 100%; overflow: hidden; background: #ffffff; }
+            #editor { width: 100%; height: 100%; }
+            #status { 
+                position: fixed; 
+                bottom: 0; 
+                right: 0; 
+                padding: 4px 8px; 
+                background: #007acc; 
+                color: white; 
+                font-size: 12px;
+                display: none;
+            }
+            #status.connected { display: block; background: #4caf50; }
+            #status.disconnected { display: block; background: #f44336; }
+        </style>
+    </head>
+    <body>
+        <div id="editor"></div>
+        <div id="status">LSP</div>
+        
+        <script src="https://cdnjs.cloudflare.com/ajax/libs/monaco-editor/0.45.0/min/vs/loader.min.js"></script>
+        <script>
+            let editor;
+            let lspSocket;
+            let lspEnabled = false;
+            
+            require.config({ paths: { vs: 'https://cdnjs.cloudflare.com/ajax/libs/monaco-editor/0.45.0/min/vs' } });
+            require(['vs/editor/editor.main'], function () {
+                editor = monaco.editor.create(document.getElementById('editor'), {
+                    value: '',
+                    language: 'java',
+                    theme: 'vs',
+                    fontSize: 14,
+                    minimap: { enabled: true },
+                    automaticLayout: true,
+                    quickSuggestions: true,
+                    suggestOnTriggerCharacters: true
+                });
+                
+                window.webkit.messageHandlers.editorReady.postMessage('ready');
+                
+                let changeTimeout;
+                editor.onDidChangeModelContent(() => {
+                    clearTimeout(changeTimeout);
+                    changeTimeout = setTimeout(() => {
+                        window.webkit.messageHandlers.contentChanged.postMessage(editor.getValue());
+                    }, 100);
+                });
+                
+                editor.focus();
+            });
+            
+            function connectLSP() {
+                const statusEl = document.getElementById('status');
+                statusEl.className = 'disconnected';
+                statusEl.textContent = 'LSP Connecting...';
+                
+                lspSocket = new WebSocket('ws://localhost:8080');
+                
+                lspSocket.onopen = () => {
+                    lspEnabled = true;
+                    statusEl.className = 'connected';
+                    statusEl.textContent = 'LSP Connected';
+                    
+                    lspSocket.send(JSON.stringify({
+                        jsonrpc: '2.0',
+                        id: 1,
+                        method: 'initialize',
+                        params: {
+                            processId: null,
+                            rootUri: 'file:///workspace',
+                            capabilities: {}
+                        }
+                    }));
+                };
+                
+                lspSocket.onclose = () => {
+                    lspEnabled = false;
+                    statusEl.className = 'disconnected';
+                    statusEl.textContent = 'LSP Disconnected';
+                };
+                
+                lspSocket.onerror = () => {
+                    statusEl.className = 'disconnected';
+                    statusEl.textContent = 'LSP Error';
+                };
+            }
+            
+            function setContent(content, language) {
+                if (editor) {
+                    monaco.editor.setModelLanguage(editor.getModel(), language || 'java');
+                    editor.setValue(content);
+                }
+            }
+            
+            function enableLSP() {
+                connectLSP();
+            }
+        </script>
+    </body>
+    </html>
+    """
     
     // Embedded Monaco HTML as fallback
     static let monacoHTML = """
